lnwallet: move commitChain to its own file.
